--- breadth-first-traversal/breadth-first-traversal.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.329
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "breadth-first-traversal", "breadth-first-traversal\breadth-first-traversal.csproj", "{CFB9575D-2337-45B2-890E-810BD84D5C6C}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnitTests", "UnitTests\UnitTests.csproj", "{31E2EA09-D2E6-428C-BACC-C0CB5A2F3B03}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{CFB9575D-2337-45B2-890E-810BD84D5C6C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CFB9575D-2337-45B2-890E-810BD84D5C6C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CFB9575D-2337-45B2-890E-810BD84D5C6C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CFB9575D-2337-45B2-890E-810BD84D5C6C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{31E2EA09-D2E6-428C-BACC-C0CB5A2F3B03}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{31E2EA09-D2E6-428C-BACC-C0CB5A2F3B03}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{31E2EA09-D2E6-428C-BACC-C0CB5A2F3B03}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{31E2EA09-D2E6-428C-BACC-C0CB5A2F3B03}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {F8367902-91D0-4ECF-B7C6-A363848633D0}
-	EndGlobalSection
-EndGlobal
