--- breadth-first-traversal/breadth-first-traversal/breadth-first-traversal.csproj
@@ -1,9 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <RootNamespace>breadth_first_traversal</RootNamespace>
-  </PropertyGroup>
-
-</Project>
