+++ challenges/breadth-first-traversal/assets/program-run.PNG
