+++ challenges/breadth-first-traversal/assets/whiteboard.jpg
