+++ challenges/breadth-first-traversal/breadth-first-traversal/Classes/Node.cs
