+++ challenges/breadth-first-traversal/breadth-first-traversal/Classes/Tree.cs
